chore(marketplace): rename catalog entities and processors
