+++ workspaces/marketplace/examples/all-orgs.yaml
@@ -1,0 +1,8 @@
+apiVersion: backstage.io/v1alpha1
+kind: Location
+metadata:
+  name: all-orgs
+  description: A collection of all the marketplace orgs
+spec:
+  targets:
+    - ./orgs/redhat.yaml
