+++ workspaces/marketplace/examples/all-plugins.yaml
@@ -1,0 +1,13 @@
+apiVersion: backstage.io/v1alpha1
+kind: Location
+metadata:
+  name: all-plugins
+  description: A collection of all the marketplace plugins
+spec:
+  targets:
+    - ./plugins/3scale.yaml
+    - ./plugins/acr.yaml
+    - ./plugins/keycloak.yaml
+    - ./plugins/quay.yaml
+    - ./plugins/tekton.yaml
+    - ./plugins/topology.yaml
