--- workspaces/marketplace/examples/entries/all.yaml
@@ -1,15 +1,0 @@
-apiVersion: backstage.io/v1alpha1
-kind: Location
-metadata:
-  name: entries
-  description: A collection of all the catalog entities
-spec:
-  targets:
-    - ./group.yaml
-    - ./pluginList.yaml
-    - ./3scale/metadata.yaml
-    - ./acr/metadata.yaml
-    - ./keycloak/metadata.yaml
-    - ./quay/metadata.yaml
-    - ./tekton/metadata.yaml
-    - ./topology/metadata.yaml
