+++ workspaces/marketplace/examples/plugins/3scale.yaml
