+++ workspaces/marketplace/examples/plugins/acr.yaml
