+++ workspaces/marketplace/examples/plugins/keycloak.yaml
