+++ workspaces/marketplace/examples/plugins/quay.yaml
