--- workspaces/marketplace/examples/template/content/catalog-info.yaml
@@ -1,8 +1,0 @@
-apiVersion: backstage.io/v1alpha1
-kind: Component
-metadata:
-  name: ${{ values.name | dump }}
-spec:
-  type: service
-  owner: user:guest
-  lifecycle: experimental
