--- workspaces/marketplace/examples/template/content/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "${{ values.name }}",
-  "private": true,
-  "dependencies": {}
-}
