+++ workspaces/marketplace/plugins/catalog-backend-module-marketplace/src/module.ts
@@ -18,8 +18,8 @@ import {
   createBackendModule,
 } from '@backstage/backend-plugin-api';
 import { catalogProcessingExtensionPoint } from '@backstage/plugin-catalog-node/alpha';
-import { MarketplacePluginProcessor } from './MarketPlaceEntityProcessor';
-import { MarketplacePluginListProcessor } from './MarketPlacePluginListProcessor';
+import { MarketplacePluginProcessor } from './MarketplacePluginProcessor';
+import { MarketplacePluginListProcessor } from './MarketplacePluginListProcessor';
 
 /**
  * @public
